+++ packages/jest-config/src/normalize.ts
@@ -769,9 +769,12 @@ export default async function normalize(
       case 'moduleDirectories':
       case 'testMatch':
         {
+          const option = oldOptions[key];
+          const rawValue =
+            Array.isArray(option) || option == null ? option : [option];
           const replacedRootDirTags = _replaceRootDirTags(
             escapeGlobCharacters(options.rootDir),
-            oldOptions[key],
+            rawValue,
           );
 
           if (replacedRootDirTags) {